backend/phonehub/requirements.txt
Django>=5.1
psycopg2-binary>=2.9
python-dotenv>=1.0.1
gunicorn>=23.0.0
strawberry-graphql-django>=0.58.0
strawberry-django-auth>=0.378.4
django-choices-field>=2.3.0
cryptography>=44.0.2
django-code-generator